requirements.txt
python-mnist
numpy~=1.19.5
matplotlib~=3.3.3